Move `FarmerProviderRecordProcessor` into its own module in farmer library
